requirements.txt
Flask>=2.0.2
Flask_SocketIO>=5.1.1
python-dotenv>=0.20.0
